Cleanup: Move ObjectFile into test project
