+++ common/test/CMakeLists.txt
@@ -9,6 +9,7 @@ set(COMMON_TEST_UTILS_SOURCE
         "${COMMON_TEST_SOURCE_DIR}/catch/Matchers.h"
         "${COMMON_TEST_SOURCE_DIR}/catch/StringMakers.cpp"
         "${COMMON_TEST_SOURCE_DIR}/catch/StringMakers.h"
+        "${COMMON_TEST_SOURCE_DIR}/io/ObjectFile.h"
         "${COMMON_TEST_SOURCE_DIR}/io/TestEnvironment.cpp"
         "${COMMON_TEST_SOURCE_DIR}/io/TestEnvironment.h"
         "${COMMON_TEST_SOURCE_DIR}/io/TestFileSystem.cpp"
